Autobot/requirements.txt
selenium==4.10.0
PyQt5==5.15.9
pyautogui==0.9.54
python-dotenv==1.0.1